feat(controllers): a logic editor on the controller's own page
A new Logic tab where a program is built from steps rather than bytes.
Values get a name, a kind and what they are wired to; steps are IF, set a
value, timers, counters, PID loops and latches, with the whole instruction
set reachable through them. Nothing in the editor mentions a stack, which
is the point.

Tag binding reads the controller's real points when the tab opens, so
connecting a value to a point is a list rather than an id typed from
memory, and the system metrics are offered by name.

Checking before sending is not politeness. A controller that refuses a
program says so on a serial console and nowhere else - over REST a rejected
program reports exactly what no program at all reports - so an operator who
sends something broken would otherwise see a successful upload, a cheerful
"staged", and then nothing happening, forever. Everything is checked on the
platform first against this device's own profile and points, and Send
refuses to write what would not run.

Also widens the program id to the u32 it is on the wire. Held as an int it
was refused by the JSON parser before the compiler ever saw it, and with a
500 rather than anything an operator could act on; 0xDEADBEEF now
round-trips, with a test.

Verified against the bench board end to end: blocks posted as JSON,
type-checked, compiled, accepted by the device's own verifier and staged.

+++ application/src/main/java/org/thingsboard/server/service/inferrix/ilb/IlbBlock.java
@@ -43,7 +43,7 @@ public final class IlbBlock {
     private IlbBlock() {}
 
     /** A whole program, as posted by the editor. */
-    public record Program(int programId, long programVersion, int profile, long scanPeriodMs,
+    public record Program(long programId, long programVersion, int profile, long scanPeriodMs,
                           List<Tag> tags, List<Statement> statements) {}
 
     /**
